requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-auth>=2.23.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
anthropic>=0.39.0
